:boom: recreate app -issue #21
