--- src/components/Header/Header.module.css
@@ -1,13 +1,0 @@
-.header {
-  display: flex;
-  justify-content: center;
-  column-gap: 3em;
-}
-
-.header a {
-  transition: opacity .4s;
-}
-
-.header a:hover {
-  opacity: .7;
-}
