--- src/components/Main/Main.module.css
@@ -1,13 +1,0 @@
-.main {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  padding: 6rem;
-  min-height: 100vh;
-}
-
-.code {
-  font-weight: 700;
-  font-family: var(--font-mono);
-}
