+++ src/pages/_app.js
@@ -1,0 +1,5 @@
+import '@/styles/globals.scss'
+
+export default function App({ Component, pageProps }) {
+  return <Component {...pageProps} />
+}
