--- src/pages/index.jsx
@@ -1,22 +1,0 @@
-import Head from 'next/head';
-import { Main } from '@/components/Main';
-import { useBackground } from '@/hooks/useBackground';
-
-export default function Home() {
-
-  useBackground();
-
-  return (
-    <>
-      <Head>
-        <title>shun kusakabe | portfolio site</title>
-        <meta
-          name="description"
-          content="This is Shun's portfolio site. Generated by create next app"
-        />
-      </Head>
-
-      <Main title="index" />
-    </>
-  );
-}
