--- src/pages/works.jsx
@@ -1,18 +1,0 @@
-import Head from 'next/head';
-import { Main } from '@/components/Main';
-
-export default function Home() {
-  return (
-    <>
-      <Head>
-        <title>Works | shun kusakabe | portfolio site</title>
-        <meta
-          name="description"
-          content="This is Shun's Works page. Generated by create next app"
-        />
-      </Head>
-
-      <Main title="works" />
-    </>
-  );
-}
