--- src/styles/Home.module.css
@@ -1,6 +1,0 @@
-/* Mobile */
-@media (max-width: 700px) {
-  .content {
-    padding: 4rem;
-  }
-}
